work on eliminating NMS usage
