+++ BuildVivecraftSpigotExtensions/build.gradle.kts
@@ -61,6 +61,6 @@ tasks.named("assemble").configure {
 tasks {
     compileJava {
         options.encoding = Charsets.UTF_8.name() // We want UTF-8 for everything
-        options.release.set(8)
+        options.release.set(16)
     }
 }
