+++ Compatibility/Vivecraft_1_19_R1/src/main/java/org/vivecraft/compatibility/Vivecraft_1_19_R1.java
@@ -1,8 +1,19 @@
 package org.vivecraft.compatibility;
 
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelInboundHandlerAdapter;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
+import net.minecraft.network.Connection;
+import net.minecraft.network.protocol.game.ServerboundPlayerActionPacket;
+import net.minecraft.network.protocol.game.ServerboundUseItemOnPacket;
+import net.minecraft.network.protocol.game.ServerboundUseItemPacket;
+import net.minecraft.network.syncher.EntityDataAccessor;
+import net.minecraft.network.syncher.SynchedEntityData;
+import net.minecraft.server.network.ServerGamePacketListenerImpl;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.Pose;
 import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.entity.ai.goal.SwellGoal;
 import net.minecraft.world.entity.ai.goal.WrappedGoal;
@@ -17,6 +28,7 @@ import net.minecraft.world.phys.Vec3;
 import org.bukkit.Location;
 import org.bukkit.craftbukkit.v1_19_R1.entity.CraftCreeper;
 import org.bukkit.craftbukkit.v1_19_R1.entity.CraftEnderman;
+import org.bukkit.craftbukkit.v1_19_R1.entity.CraftPlayer;
 import org.bukkit.entity.Creeper;
 import org.bukkit.entity.Enderman;
 import org.bukkit.entity.Player;
@@ -24,15 +36,18 @@ import org.bukkit.util.Vector;
 import org.vivecraft.Reflector;
 import org.vivecraft.VSE;
 import org.vivecraft.VivePlayer;
+import org.vivecraft.utils.AimFixHandler;
+import org.vivecraft.utils.Vector3;
 
 import java.util.AbstractCollection;
 import java.util.EnumSet;
+import java.util.UUID;
 import java.util.function.Predicate;
 
 public class Vivecraft_1_19_R1 implements VivecraftCompatibility {
 
     @Override
-    public void editCreeper(Creeper creeper, double radius) {
+    public void injectCreeper(Creeper creeper, double radius) {
         net.minecraft.world.entity.monster.Creeper e = ((CraftCreeper) creeper).getHandle();
         AbstractCollection<WrappedGoal> goalB = (AbstractCollection<WrappedGoal>) Reflector.getFieldValue(Reflector.availableGoals, ((Mob) e).goalSelector);
         for (WrappedGoal b : goalB) {
@@ -45,7 +60,7 @@ public class Vivecraft_1_19_R1 implements VivecraftCompatibility {
     }
 
     @Override
-    public void editEnderman(Enderman enderman) {
+    public void injectEnderman(Enderman enderman) {
         EnderMan e = ((CraftEnderman) enderman).getHandle();
         AbstractCollection<WrappedGoal> targets = (AbstractCollection<WrappedGoal>) Reflector.getFieldValue(Reflector.availableGoals, ((Mob) e).targetSelector);
         for (WrappedGoal b : targets) {
@@ -65,6 +80,125 @@ public class Vivecraft_1_19_R1 implements VivecraftCompatibility {
         }
         e.goalSelector.addGoal(1, new CustomGoalStare(e));
     }
+
+    public void injectPlayer(Player player) {
+        EntityDataAccessor<Pose> poseObj = (EntityDataAccessor<Pose>) Reflector.getFieldValue(Reflector.Entity_Data_Pose, player);
+        SynchedEntityData dataWatcher = ((CraftPlayer) player).getHandle().getEntityData();
+        Int2ObjectOpenHashMap<SynchedEntityData.DataItem<?>> entries = (Int2ObjectOpenHashMap<SynchedEntityData.DataItem<?>>) Reflector.getFieldValue(Reflector.SynchedEntityData_itemsById, dataWatcher);
+        InjectedDataWatcherItem item = new InjectedDataWatcherItem(poseObj, Pose.STANDING, player);
+        entries.put(poseObj.getId(), item);
+
+        Connection netManager = ((CraftPlayer) player).getHandle().connection.connection;
+        netManager.channel.pipeline().addBefore("packet_handler", "vr_aim_fix", new AimFixHandler(netManager));
+    }
+
+
+    public class AimFixHandler extends ChannelInboundHandlerAdapter {
+        private final Connection netManager;
+
+        public AimFixHandler(Connection netManager) {
+            this.netManager = netManager;
+        }
+
+        @Override
+        public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
+            net.minecraft.world.entity.player.Player player = ((ServerGamePacketListenerImpl) netManager.getPacketListener()).player;
+            boolean isCapturedPacket = msg instanceof ServerboundUseItemPacket || msg instanceof ServerboundUseItemOnPacket || msg instanceof ServerboundPlayerActionPacket;
+            UUID uuid = player.getGameProfile().getId();
+            if (!VSE.vivePlayers.containsKey(uuid) || !VSE.vivePlayers.get(uuid).isVR() || !isCapturedPacket || player.getServer() == null) {
+                // we don't need to handle this packet, just defer to the next handler in the pipeline
+                ctx.fireChannelRead(msg);
+                return;
+            }
+
+            player.getServer().submit(() -> {
+                // Save all the current orientation data
+                Vec3 oldPos = player.position();
+                Vec3 oldPrevPos = new Vec3(player.xo, player.yo, player.zo);
+                float oldPitch = player.getXRot();
+                float oldYaw = player.getYRot();
+                float oldYawHead = player.yHeadRot; // field_70759_as
+                float oldPrevPitch = player.xRotO;
+                float oldPrevYaw = player.yRotO;
+                float oldPrevYawHead = player.yHeadRotO; // field_70758_at
+                float oldEyeHeight = player.getEyeHeight();
+
+                VivePlayer data = null;
+                if (VSE.vivePlayers.containsKey(uuid) && VSE.vivePlayers.get(uuid).isVR()) { // Check again in case of race condition
+                    data = VSE.vivePlayers.get(uuid);
+                    Location pos = data.getControllerPos(0);
+                    Vector aim = data.getControllerDir(0);
+
+                    // Inject our custom orientation data
+                    player.setPosRaw(pos.getX(), pos.getY(), pos.getZ());
+                    player.xo = pos.getX();
+                    player.yo = pos.getY();
+                    player.zo = pos.getZ();
+                    player.setXRot((float) Math.toDegrees(Math.asin(-aim.getY())));
+                    player.setYRot((float) Math.toDegrees(Math.atan2(-aim.getX(), aim.getZ())));
+                    player.xRotO = player.getXRot();
+                    player.yRotO = player.yHeadRotO = player.yHeadRot = player.getYRot();
+                    Reflector.setFieldValue(Reflector.Entity_eyeHeight, player, 0);
+
+                    // Set up offset to fix relative positions
+                    // P.S. Spigot mappings are stupid
+                    data.offset = oldPos.add(-pos.getX(), -pos.getY(), -pos.getZ());
+                }
+
+                // Call the packet handler directly
+                // This is several implementation details that we have to replicate
+                try {
+                    if (netManager.isConnected()) {
+                        try {
+                            ((Packet) msg).handle(this.netManager.getPacketListener());
+                        } catch (RunningOnDifferentThreadException runningondifferentthreadexception) {
+                        }
+                    }
+                } finally {
+                    // Vanilla uses SimpleInboundChannelHandler, which automatically releases
+                    // by default, so we're expected to release the packet once we're done.
+                    ReferenceCountUtil.release(msg);
+                }
+
+                // Restore the original orientation data
+                player.setPosRaw(oldPos.x, oldPos.y, oldPos.z);
+                player.xo = oldPrevPos.x;
+                player.yo = oldPrevPos.y;
+                player.zo = oldPrevPos.z;
+                player.setXRot(oldPitch);
+                player.setYRot(oldYaw);
+                player.yHeadRot = oldYawHead;
+                player.xRotO = oldPrevPitch;
+                player.yRotO = oldPrevYaw;
+                player.yHeadRotO = oldPrevYawHead;
+                Reflector.setFieldValue(Reflector.Entity_eyeHeight, player, oldEyeHeight);
+
+                // Reset offset
+                if (data != null)
+                    data.offset = new Vector3(0, 0, 0);
+            });
+        }
+    }
+
+
+    public static class InjectedDataWatcherItem extends SynchedEntityData.DataItem<Pose> {
+        protected final Player player;
+
+        public InjectedDataWatcherItem(EntityDataAccessor<Pose> datawatcherobject, Pose t0, Player player) {
+            super(datawatcherobject, t0);
+            this.player = player;
+        }
+
+        @Override
+        public void setValue(Pose pose) {
+            VivePlayer vp = VSE.vivePlayers.get(player.getUniqueId());
+            if (vp != null && vp.isVR() && vp.crawling)
+                super.setValue(Pose.SWIMMING);
+            else
+                super.setValue(pose);
+        }
+    }
+
 
     public static class CustomGoalSwell extends SwellGoal {
