+++ src/main/java/org/vivecraft/compatibility/VivecraftCompatibility.java
@@ -2,10 +2,13 @@ package org.vivecraft.compatibility;
 
 import org.bukkit.entity.Creeper;
 import org.bukkit.entity.Enderman;
+import org.bukkit.entity.Player;
 
 public interface VivecraftCompatibility {
 
-    void editCreeper(Creeper creeper, double radius);
+    void injectCreeper(Creeper creeper, double radius);
 
-    void editEnderman(Enderman enderman);
+    void injectEnderman(Enderman enderman);
+
+    void injectPlayer(Player player);
 }
