+++ src/main/java/org/vivecraft/listeners/VivecraftCombatListener.java
@@ -1,10 +1,6 @@
 package org.vivecraft.listeners;
 
-import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.world.phys.Vec3;
 import org.bukkit.Location;
-import org.bukkit.craftbukkit.v1_19_R1.entity.CraftEntity;
-import org.bukkit.craftbukkit.v1_19_R1.entity.CraftPlayer;
 import org.bukkit.entity.*;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -15,7 +11,6 @@ import org.bukkit.util.Vector;
 import org.vivecraft.VSE;
 import org.vivecraft.VivePlayer;
 import org.vivecraft.utils.Headshot;
-
 
 public class VivecraftCombatListener implements Listener {
 
@@ -43,10 +38,8 @@ public class VivecraftCombatListener implements Listener {
             vse.getLogger().warning(" Error on projectile launch!");
         }
 
-        ServerPlayer nsme = ((CraftPlayer) pl).getHandle();
-
         Location pos = vp.getControllerPos(vp.activeHand);
-        Vec3 aim = vp.getControllerDir(vp.activeHand);
+        Vector aim = vp.getControllerDir(vp.activeHand);
 
         //this only works if the incoming speed is at max (based! on draw time)
         //TODO: properly scale in all cases.
@@ -55,16 +48,15 @@ public class VivecraftCombatListener implements Listener {
             proj.setVelocity(proj.getVelocity().multiply(vp.getDraw()));
             if (!vp.isSeated()) { //standing
                 pos = vp.getControllerPos(0);
-                Vector m = (vp.getControllerPos(1).subtract(vp.getControllerPos(0))).toVector();
-                m = m.normalize();
-                aim = new Vec3(m.getX(), m.getY(), m.getZ());
+                aim = (vp.getControllerPos(1).subtract(vp.getControllerPos(0))).toVector();
+                aim.normalize();
             }
         }
 
-        Location loc = new Location(proj.getWorld(), pos.getX() + aim.x * 0.6f, pos.getY() + aim.y * 0.6f, pos.getZ() + aim.z * 0.6f);
+        Location loc = new Location(proj.getWorld(), pos.getX() + aim.getX() * 0.6f, pos.getY() + aim.getY() * 0.6f, pos.getZ() + aim.getZ() * 0.6f);
         double velo = proj.getVelocity().length();
         proj.teleport(loc); //paper sets velocity to 0 on teleport.
-        proj.setVelocity(new Vector(aim.x * velo, aim.y * velo, aim.z * velo));
+        proj.setVelocity(new Vector(aim.getX() * velo, aim.getY() * velo, aim.getZ() * velo));
     }
 
     @EventHandler(ignoreCancelled = true, priority = EventPriority.LOW)
@@ -140,14 +132,13 @@ public class VivecraftCombatListener implements Listener {
             } else if (damaged instanceof Fireball) {
                 VivePlayer vp = VSE.vivePlayers.get(damager.getUniqueId());
                 if (vp != null && vp.isVR()) {
-                    Vec3 dir = vp.getHMDDir();
+                    Vector dir = vp.getHMDDir();
                     //Interesting experiment.
                     //We know the player's look is read immediately after this event returns.
                     //Override it here. It should be set back to normal next tick.
                     //And ideally nothing weird happens because of it.
 
-                    ((CraftEntity) damager).getHandle().setXRot((float) Math.toDegrees(Math.asin(dir.y / dir.length())));
-                    ((CraftEntity) damager).getHandle().setYRot((float) Math.toDegrees(Math.atan2(-dir.x, dir.z)));
+                    damager.setRotation((float) Math.toDegrees(Math.asin(dir.getY() / dir.length())), (float) Math.toDegrees(Math.atan2(-dir.getX(), dir.getY())));
                 }
             }
         }
