+++ src/main/java/org/vivecraft/listeners/VivecraftItemListener.java
@@ -1,7 +1,5 @@
 package org.vivecraft.listeners;
 
-import net.minecraft.util.Mth;
-import net.minecraft.world.phys.Vec3;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -13,11 +11,6 @@ import org.vivecraft.VSE;
 import org.vivecraft.VivePlayer;
 
 public class VivecraftItemListener implements Listener {
-    VSE vse = null;
-
-    public VivecraftItemListener(VSE vse) {
-        this.vse = vse;
-    }
 
     @EventHandler(ignoreCancelled = true, priority = EventPriority.LOW)
     public void onPlayerDropItem(PlayerDropItemEvent event) {
@@ -35,12 +28,12 @@ public class VivecraftItemListener implements Listener {
             Vector v = new Vector();
             float yaw = player.getLocation().getYaw();
             float pitch = -player.getLocation().getPitch();
-            v.setX((double) (-Mth.sin(yaw * 0.017453292F) * Mth.cos(player.getLocation().getPitch() * 0.017453292F) * f2));
-            v.setZ((double) (Mth.cos(yaw * 0.017453292F) * Mth.cos(player.getLocation().getPitch() * 0.017453292F) * f2));
-            v.setY((double) (Mth.sin(pitch * 0.017453292F) * f2 + 0.1F));
+            v.setX(-Math.sin(yaw * 0.017453292F) * Math.cos(player.getLocation().getPitch() * 0.017453292F) * f2);
+            v.setZ(Math.cos(yaw * 0.017453292F) * Math.cos(player.getLocation().getPitch() * 0.017453292F) * f2);
+            v.setY(Math.sin(pitch * 0.017453292F) * f2 + 0.1F);
 
-            Vec3 aim = vp.getControllerDir(0);
-            event.getItemDrop().teleport(vp.getControllerPos(0).add(0.2f * aim.x, 0.25f * aim.y - 0.2f, 0.2f * aim.z));
+            Vector aim = vp.getControllerDir(0);
+            event.getItemDrop().teleport(vp.getControllerPos(0).add(0.2f * aim.getX(), 0.25f * aim.getY() - 0.2f, 0.2f * aim.getZ()));
             event.getItemDrop().setVelocity(v);
         }
     }
